pyproject.toml
[project]
name = "disinfo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.13.*"
dependencies = [
    "pillow==11.2.1",
    "watchdog>=3.0.0",
    "requests>=2.25.1",
    "redis[hiredis]>=4.5.4",
    "schedule>=1.2.0",
    "paho-mqtt>=2.1.0",
    "pydash>=7.0.3",
    "arrow>=1.2.3",
    "colour>=0.1.5",
    "typer[all]",
    "typing-extensions>=4.5.0",
    "idfm_api>=1.1.1",
    "pendulum>=2.1.2",
    "scipy>=1.17.0",
    "pydantic>=2.12.5",
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.1",
    "websockets>=11.0.3",
    "sympy>=1.12",
    "pycairo>=1.24.0",
    "suncalc>=0.1.3",
    "vectormath>=0.2.2",
    "spectra>=0.0.11",
    "cairosvg>=2.7.1",
    "pyquery>=2.0.0",
    "scikit-image>=0.22.0",
    "websocket-client>=1.8.0",
    "geopy>=2.4.1",
    "cartopy>=0.23.0",
    "shazamio>=0.6.0",
    "pyaudio>=0.2.14",
    "py-mjpeg>=1.0.1",
    "ipython",
    "matplotlib",
    "redis-om>=1.0.6",
]

[tool.uv.sources]
websocket-rpi-matrix = { path = "clients/websocket-rpi-matrix", editable = true }

[dependency-groups]
dev = [
    "watchdog>=6.0.0",
    "websocket-rpi-matrix",
]
